feat: 記事単体取得APIに画像リストを追加し、Schemaに定義していた処理をService用クラスに分離
